cast auf unsigned damit es kompiliert

+++ Thermostaten/src/Bussi.cpp
@@ -33,7 +33,7 @@ void Bussi::propagate() {
 	auto expfactor = exp(-m_dtime / couplingtime);
 
 	auto rate = (1.0 - expfactor)*temperature_ratio / m_poly.monomers.size();
-	double chisquared = Rand::real_chisquared(m_poly.monomers.size() - 1);
+	double chisquared = Rand::real_chisquared( (unsigned) m_poly.monomers.size() - 1);
 	//for (unsigned i = 1; i < m_poly.monomers.size(); i++){ // n-1 times
 	//	chisquared += pow(Rand::real_normal(), 2);
 	//}
